+++ SkeppOHoj/Models/InsuranceClaim.cs
@@ -2,8 +2,8 @@
 {
     public class InsuranceClaim
     {
-        public long InsuranceClaimId { get; set; }
-        public long StatusId { get; set; }
+        public int InsuranceClaimId { get; set; }
+        public int StatusId { get; set; }
         public DateTime Date { get; set; }
         public string? Description{ get; set; }
         public double Amount { get; set; }
