+++ SkeppOHoj/Models/InsuranceClaimComment.cs
@@ -2,8 +2,8 @@
 {
     public class InsuranceClaimComment
     {
-        public long InsuranceClaimCommentId { get; set; }
-        public long InsuranceClaimId { get; set; }
+        public int InsuranceClaimCommentId { get; set; }
+        public int InsuranceClaimId { get; set; }
         public string? Comment { get; set; }
         public DateTime Date { get; set; }
     }
